Update pre-commit config with new versions and fix
I discovered errors when calling `pre-commit run eslint --all-files --verbose`. It appears that I nee d to be in the frontend directory before executing `npx eslint —fix`. I also was using the wrong argument for filtering types.

Also, ran `pre-commit autoupdate` and updated the revisions to the lastest revs of flake8 and black.