Add Django unit test for successful photo uploads
Implementing this test required research and lots of additional imports. The test is wrapped in an annotation that overrides where Django’s media directory is located to a temporary location so it can be deleted after the test finishes. The test creates an image and sends a POST request to the photo_upload view. It then checks the response code and whether the photo’s original file name exists in the Photo table.