Add python-decouple
We need to prepare settings.py for use in production. To do this we will use `python-decouple` to help create separate .env and .env.prod files with development and production-ready settings (e.g., DEBUG=False, a different SECRET_KEY, etc.) and adjust your settings file to load the appropriate file based on an environment variable.