Update docker-compose with backend mount for .env
We’re no longer using the `env_file` directive and, instead, mounting the .env file as a volume. This will allow settings.py to read the .env file in a future commit. Reading the .env file from inside the container will allow python-decouple to read it and allow us to switch between production and development environment variables.