Update REAMDE for Deploying and Publishing Images
Add an instruction to setup the development environment before attempting to deploy.

Update command usage (`docker compose` is preferred to `docker-compose`).

In the Publishing Docker Images section, remove unnecessary versioning steps and replace the build process with one that outputs and pushes images for AMD64 Linux and Apple Silicon. Make explicit the important of setting a `VITE_API_URL` environment variable before running the build command for the frontend.