Update .env example files
Now we have a DEVELOPMENT flag to decouple the database setting from the DEBUG setting. We also app the `sharedmemories` hostname that the proxy service is using to the ALLOWED_HOSTS environment variable. Finally, we attempt to standardize on relevative VITE_API_URL paths when deploying with the proxy service/docker.